%%  WARNING: This file was automatically generated from the associated 
%%  fra_data.mid file. Do NOT edit this Rd file to make a change. Instead, 
%%  edit the fra_data.mid file in the project MID directory. Once the 
%%  fra_data.mid file has been updated, this Rd file, and all other 
%%  documentation (such as corresponding LaTeX, SGML and HTML documentation) 
%%  should be regenerated using the mid.pl Perl script. 

%% R documentation for the ecgrr function

\name{ecgrr}
\alias{ecgrr}
\title{Electrocardiogram R-R Interval Data}
\concept{measured time series}
\description{These data are from the file \code{rr1.txt}
in the `RR interval time series modeling:
A challenge from PhysioNet and Computers in Cardiology 2002' site of PhysioNet.
sponsored by NIH's National Center for Research Resources.

The data are the RR intervals (beat-to-beat intervals measured
between successive peaks of the QRS complex) for patients in normal sinus rhythm
(record 16265 of the MIT-BIH database).}

\seealso{
\code{\link{beamchaos}}, \code{\link{eegduke}}, \code{\link{lorenz}}, \code{\link{pd5si}}.}

\examples{
plot(ecgrr)
}
\keyword{data}

