%%  WARNING: This file was automatically generated from the associated 
%%  fra_data.mid file. Do NOT edit this Rd file to make a change. Instead, 
%%  edit the fra_data.mid file in the project MID directory. Once the 
%%  fra_data.mid file has been updated, this Rd file, and all other 
%%  documentation (such as corresponding LaTeX, SGML and HTML documentation) 
%%  should be regenerated using the mid.pl Perl script. 

%% R documentation for the eegduke function

\name{eegduke}
\alias{eegduke}
\title{Electroencephalogram Recordings of a Seizure}
\description{EEG of a patient undergoing ECT therapy for clinical depression at the ECT Lab at Duke.
The data are fluctuations in electrical potential at a point on the patient's scalp
during seizure, one of several recorded channels. They are measured in microvolts and
represent measurements taken at time intervals of roughly \eqn{1/40} of a second.}

\seealso{
\code{\link{beamchaos}}, \code{\link{ecgrr}}, \code{\link{lorenz}}, \code{\link{pd5si}}.}

\examples{
plot(eegduke)
}
\keyword{data}

