% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal-package.R
\docType{package}
\name{fractal-package}
\alias{fractal-package}
\alias{targeted}
\title{Fractal visualization}
\description{
Visualization of Julia and Mandelbrot sets.
}
\author{
Klaus K. Holst Maintainer: <klaus@holst.it>
}
\keyword{package}
