% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractalr-package.R
\docType{package}
\name{fractalr-package}
\alias{fractalr-package}
\alias{fractalr}
\title{Fractal visualization}
\description{
Visualization of Julia and Mandelbrot sets.
}
\examples{
if (interactive()) fractalr::ui()
}
\author{
Klaus K. Holst Maintainer: <klaus@holst.it>
}
\keyword{package}
