\name{kochPattern}
\alias{kochPattern}
\title{
 Define the pattern used to generate General Koch Curve.
}
\description{
 Allow the user to define a pattern to generate Koch Curve by clicking on a plot. Returns a dataframe with 2 columns: first one is the distance from the pattern point to original point. second one is the angle between horizontal axis and the segment connecting the origin and the pattern point.
}
\usage{
kochPattern()
}

\examples{

#ptn = kochPattern()

}
