% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{Plot chaotic logistic growth}
\usage{
logistic(
  x.init = 0.01,
  gen = 50,
  rlim = c(3.2, 4),
  ylim = c(0, 1),
  pch = ".",
  col = rgb(0, 0, 0, 0.3),
  div = 2000,
  k = 32
)
}
\arguments{
\item{x.init}{numeric, the initial population size, X(0)}

\item{gen}{numeric, the number of generations}

\item{rlim}{numeric, the growth rate, r, range, for plotting}

\item{ylim}{numeric, the y-axis limits, for plotting}

\item{pch}{the plotting character for points}

\item{col}{the plotting color for points}

\item{div}{numeric, the number of divisions along the x-axis, for plotting}

\item{k}{numeric, the number of burn-out generations

Consider a population with growth

X(t+1) = r * X(t) * (1 - X(t))

where r is the growth rate and X(t) is the population size at generation t, with
0 <= X(t) <= 1. This equation shows chaotic behaviour for r > 3.57.
For example, for r < 3, the population growths until converging to a limit. For
3 < r < 3.57 the population oscillates between two values. Beyond r = 3.57, the
behaviour becomes chaotic.

This function plots the long-term values of X(t) as a function of r.}
}
\description{
Plot chaotic logistic growth
}
\examples{
logistic()

\dontrun{
# higher resolution but slow
logistic(div=4000, k=128)
}

}
\references{
Tien-Yin Li and James A. Yorke. (1975) Period three implies chaos. \emph{
The American Mathematical Monthly}, 82: 985--992.

May R. (1976) Simple mathematical models with very complicated dynamics.
\emph{Nature}, 261: 459--467.
}
\author{
Mario dos Reis
}
