\name{plot.BarnsleyFern}
\alias{plot.BarnsleyFern}

\title{
S3 Method to Plot a BarnsleyFern Object
}
\description{
Display a BarnsleyFern object in the R console
}
\usage{
plot.BarnsleyFern(x, ...)
}
\arguments{
  \item{x}{
A BarnsleyFern Object.
}
  \item{\dots}{
see in plot() function
}
}

\examples{
## b = barnsleyFern()
## plot(b)
}