\name{plot.Julia}
\alias{plot.Julia}
\title{
S3 Method to Plot a Julia Object
}
\description{
Display a Julia object in the R console
}
\usage{
\method{plot}{Julia}(x,...)
}
\arguments{
  \item{x}{
A matrix object,typically with class of Julia .
}
  \item{\dots}{
see in plot() function
}
} 

\examples{
#j = julia(0.0284+ 0.4281i)
#plot(j)
}
