\name{plot.KochPlusCurve}
\alias{plot.KochPlusCurve}
\title{
S3 Method for a KochPlusCurve Object
}
\description{
Display a KochPlusCurve object in the R console
}
\usage{
\method{plot}{KochPlusCurve}(x,...)
}
\arguments{
  \item{x}{
A matrix object with class KochPlusCurve.
}
  \item{\dots}{
see in plot() function
}
} 

\examples{
#koch = koch(0, 10, 0, 0, 1)
#plot(koch)
}
