\name{plot.KochSnowFlake}
\alias{plot.KochSnowFlake}
\title{
S3 Method for a KochSnowFlake Object
}
\description{
Display a KochSnowFlake object in the R console
}
\usage{
\method{plot}{KochSnowFlake}(x,...)
}
\arguments{
  \item{x}{
A matrix object,with class of KochSnowFlake.
}
  \item{\dots}{
see in plot() function
}
} 

\examples{
#snow = kochSnowFlake(2, 3)
#plot(snow)
}
