\name{plot.Mandelbrot}
\alias{plot.Mandelbrot}
\title{
S3 Method to Plot a Julia Object
}
\description{
Display a Mandelbrot object in the R console
}
\usage{
\method{plot}{Mandelbrot}(x,...)
}
\arguments{
  \item{x}{
A matrix object,typically with class of Mandelbrot.
}
  \item{\dots}{
see in plot() function
}
} 

\examples{
#m = mandelbrot()
#plot(m)
}
