% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\title{User-interface for visualizing fractals}
\usage{
ui(...)
}
\arguments{
\item{...}{additional arguments to lower level functions}
}
\description{
User-interface for visualizing fractals
}
\examples{
if (interactive()) fractalr::ui()
}
