% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoom}
\alias{zoom}
\title{Zoom into the Mandelbrot set}
\usage{
zoom(
  col = c(heat.colors(49), "black"),
  fun = "mandelbrot",
  mu,
  nx = 600,
  ny = 600,
  iter = NULL,
  plot = TRUE,
  invisible = TRUE
)
}
\arguments{
\item{col}{the color palette to plot the set}

\item{fun}{which fractal function to plot, currently only the Mandelbrot}

\item{mu}{parameter from the Julia set (not used)}

\item{nx}{number of subdivisions in the x-axis (for plotting)}

\item{ny}{number of subdivisions in the y-axis (for plotting)}

\item{iter}{number of iterations to calculate the set}

\item{plot}{whether to plot the zoomed-in area}

\item{invisible}{whether to return a matrix with the calculated iterations}
}
\description{
Selects a region of a fractal image and zooms into it.
}
\details{
Currently the function only zooms into the Mandelbrot set. The set should
be plotted with the \code{image} function. Then, a call to \code{zoom} uses
\code{locator} to read the coordinates of two points on the image, selected
by the user using the mouse.

It is recommended that you increase the number of iterations as you zoom
into the boundary areas of the set.
}
\examples{
\dontrun{
z <- mandelbrot()
# The set is black
cols <- c(hcl.colors(n=100, pal="Lisbon"), "black")
image(z, col=cols, las=1)
points(c(-.375,.125), c(.5, 1), pch=19, col="orange")

# Run the zoom command below, then use your mouse to click on
# the two orange dots on the image
zoom(col=cols, iter=100)
# now try to zoom in deeply by clicking on two arbitrary points
# around a region of interest
}

}
