% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandelc.R
\name{mandelbrot}
\alias{mandelbrot}
\title{The Mandelbrot Set}
\usage{
mandelbrot(x = c(-3, 1), y = c(-1.8, 1.8), nx = 600, ny = 600,
  iter = 20)
}
\arguments{
\item{x}{x limits}

\item{y}{y limits}

\item{nx}{x resolution}

\item{ny}{y resolution}

\item{iter}{maximum number of iterations}
}
\value{
A list with components: \code{x}, a vector of x coordinates,
\code{y}, a vector of y coordinates, and \code{z}, a nx * ny matrix with
the set.
}
\description{
A function to calculate the Mandelbrot set
}
\examples{
z <- mandelbrot()
image(z)

}

