% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandelc.R
\name{mandelbrot}
\alias{mandelbrot}
\title{The Mandelbrot Set}
\usage{
mandelbrot(x = c(-2, 1), y = c(-1.5, 1.5), nx = 600, ny = 600,
  iter = 80)
}
\arguments{
\item{x}{x limits}

\item{y}{y limits}

\item{nx}{x resolution}

\item{ny}{y resolution}

\item{iter}{maximum number of iterations}
}
\value{
A list with components: \code{x}, a vector of x coordinates,
\code{y}, a vector of y coordinates, and \code{z}, a \code{nx} * \code{ny} matrix with
the set.
}
\description{
A function to calculate the Mandelbrot set
}
\examples{
z <- mandelbrot()
image(z, col=c(heat.colors(49), "black"), las=1)

}

